app/requirements.txt
psycopg2-binary
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask